refactor(routes): reorganize into (app)/(site)/(auth)/(dev) route groups
- Move static content pages (about, help, privacy, terms, whats-new,
  landing) to new (site) route group
- Move report + report/success into (app) — inherits MainLayout
  automatically, removes manual MainLayout imports
- Move auth callback pages (callback, auth-code-error, loading) into
  (auth)/auth/ — gets card layout; rename loading→callback-loading
- Move dev pages into new (dev) route group; consolidate /debug/badges
  under /dev/badges
- Add (site)/layout.tsx and (dev)/layout.tsx (one-liner MainLayout wrappers)
- Update internal redirect paths and cross-module imports

MainLayout is now imported in layout files only — never in pages.
