+++ src/app/(site)/layout.tsx
@@ -1,0 +1,10 @@
+import type React from "react";
+import { MainLayout } from "~/components/layout/MainLayout";
+
+export default function SiteLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}): React.JSX.Element {
+  return <MainLayout>{children}</MainLayout>;
+}
