+++ src/app/(site)/terms/page.tsx
