+++ src/app/(site)/terms/terms-content.tsx
